+++ operator.ps1
@@ -117,7 +117,10 @@ if ($IsWindows){
 . $pwshSrcDir/mdgrep_function.ps1
 . $pwshSrcDir/mdfocus_function.ps1
 
-. $pwshSrcDir/Execute-TinyTex_function.ps1
+. $pwshSrcDir/Execute-TinyTeX_function.ps1
+. $pwshSrcDir/Execute-RMarkdown_function.ps1
+. $pwshSrcDir/Inkscape-Converter_function.ps1
+
 . $pwshSrcDir/math2tex_function.ps1
 . $pwshSrcDir/tex2pdf_function.ps1
 . $pwshSrcDir/jl_function.ps1