+++ CHANGELOG.md
@@ -9,9 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - NA
 
+## [7.1.0] - 2024-02-18
+
+- Added [Execute-RMarkdown] (Alias: rmarkdown) function
+- Renamed [Execute-TinyTeX] from `Execute-TinyTex`
+- Added [Inkscape-Converter] (Alias: inkconv) function
+
 ## [7.0.0] - 2024-02-17
 
-- Added [Execute-TinyTex] (Alias: tinytex) function
+- Added [Execute-TinyTeX] (Alias: tinytex) function
 - Added [math2tex] function
 - Fixed [sleepy] PercentComplete value to 100 if it is greater than 100
 - Fixed [README.md] create command list script
@@ -676,11 +682,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [Grep-Block]: src/Grep-Block_function.ps1
 [Sort-Block]: src/Sort-Block_function.ps1
 
-[Execute-TinyTex]: src/Execute-TinyTex_function.ps1
+[Execute-TinyTeX]: src/Execute-TinyTeX_function.ps1
 [math2tex]: src/math2tex_function.ps1
+[Inkscape-Converter]: src/Inkscape-Converter_function.ps1
+
+[Execute-RMarkdown]: src/Execute-RMarkdown_function.ps1
 
 
-[unreleased]: https://github.com/btklab/posh-mocks/compare/7.0.0..HEAD
+[unreleased]: https://github.com/btklab/posh-mocks/compare/7.1.0..HEAD
+[7.1.0]: https://github.com/btklab/posh-mocks/releases/tag/7.1.0
 [7.0.0]: https://github.com/btklab/posh-mocks/releases/tag/7.0.0
 [6.10.2]: https://github.com/btklab/posh-mocks/releases/tag/6.10.2
 [6.10.1]: https://github.com/btklab/posh-mocks/releases/tag/6.10.1